Split up __main__.py
